Add extra integration specs around as: behaviour for part wrapping
